[Bronze V] Title: 이상한 기호, Time: 68 ms, Memory: 11508 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -13,10 +13,10 @@ class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         
 
-        int N = Integer.parseInt(st.nextToken());
-        int M = Integer.parseInt(st.nextToken());
+        long N = Long.parseLong(st.nextToken());
+        long M = Long.parseLong(st.nextToken());
         
-        int result = (N+M)*(N-M);
+        long result = (N+M)*(N-M);
 
         System.out.print(result);
     }
